nand-data: add "Scan (last-2)th page" and "ONFI-Cabaple" columns
There's a strange Hynix chip out there that has "(last-2)th and last page"
BBM scanning. Include a column that records this.

Also, for our "support" information (the booleans at the right side of the
table) it's useful to have an "ONFI-Capable" flag, so I added a column.

Also, the original "ONFI" column is renamed "ONFI ver." to provide
a distinction between it and the new "ONFI-Capable" column.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>