faq: ubi(fs): small spelling, URL fixups
We used `<a name=[URL]>' instead of `<a href=[URL]>' by accident.

Also, the file "faq/ubi.html" is located within the "faq" directory, so
use a simpler relative link (i.e., don't use `../faq').

Change "we have never find" to "we have never found".

Change "refer this tread" to "refer to this thread".

Change "UBIi" to "UBI", "you your" to "your".

Change "refer the" to "refer to the".

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Signed-off-by: Artem Bityutskiy <dedekind1@gmail.com>