UBI FAQ: suggest omission of vol_size option
OLPC set vol_size based on the size of the target NAND and found out
that it resulted in some systems being unbootable, where such systems
had a lot of bad blocks on their flash.

For distributors such as OLPC wishing to maximize robustness in the
face of varying bad block counts, it makes a lot of sense to avoid
the vol_size option and let it be calculated automatically.
Document this.

Signed-off-by: Daniel Drake <dsd@laptop.org>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>