Update UBIFS mounting instructions
I recently spent several days attempting to mount a UBIFS image.
Efforts were hampered by seemingly outdated documentation,
the fact that much of the documentation was for UBI images,
not UBIFS ones, and the fact that the image was for NOR memory and
all documentation I found only referenced nandsim.

After some very kind help on IRC and from Artem Bityutskiy on the
mailing list, I was able to figure things out. I think it is only
fair that I contribute back and bring the web site up to date so
others don't face the same stumbling blocks.

This patch includes the following changes:

- Acknowledge that udev _does_ now create a block device
  automatically for MTD.

- Provide instructions for mounting standalone UBIFS images as well
  as UBI images.

- Mention mtdram and its usefulness in mounting UBIFS images made
  for NOR memory.